Fix fatal error under apache/thrift 0.24
apache/thrift 0.24 added return types to the abstract TTransport methods
(isOpen(): bool, open(): void, read(int $len): string, ...). TUDPTransport
declared all six untyped, so the class could no longer be loaded at all:

  Declaration of Jaeger\Transport\TUDPTransport::isOpen() must be compatible
  with Thrift\Transport\TTransport::isOpen(): bool

That took out every request reaching Tracing::setup(), which constructs a
JaegerTransport. No code changed to cause it: thrift is an unpinned transitive
dependency (>=0.10.0 <1.0.0 here, >=0.16.0 <1.0.0 via thrift-decorators), so a
rebuild after 0.24.0 was released floated onto it.

Return types are declared but parameters are deliberately left untyped. Adding
a return type where the parent declares none is allowed; adding a parameter
type where the parent has none is not. This loads under both thrift < 0.24 and
>= 0.24, which matters because consumers are spread across 0.22, 0.23 and 0.24
and requiring 0.24 would strand the older ones.

read() can no longer fall through returning null under `: string`, so it raises
TTransportException instead. This transport only ever sends, and returning ""
would make TTransport::readAll() loop forever.

Also fixes two latent TypeErrors in the same class, both reachable by calling
Tracing::teardown() twice, which sets $alreadyTornDown at the end but never
checks it on entry:

  socket_close(): Argument #1 ($socket) must be of type Socket, null given
  socket_sendto(): Argument #1 ($socket) must be of type Socket, null given

close() is now idempotent, and flushing a closed transport raises
TTransportException, which JaegerTransport::flush() already catches and logs.

Lastly, ProbabilisticSampler used a (double) cast, deprecated as a
non-canonical cast in PHP 8.5, which the CI matrix covers.

+++ src/Jaeger/Sampler/ProbabilisticSampler.php
@@ -21,7 +21,7 @@ class ProbabilisticSampler implements Sampler
         // In practice this turns out to be 2^63 - 1 on a 64-bit system.
         $sampleMax = (int) (mt_getrandmax() << 31 | mt_getrandmax());
 
-        $this->samplingRate = (double) $samplingRate;
+        $this->samplingRate = (float) $samplingRate;
         $this->samplingBoundary = $this->samplingRate * $sampleMax;
     }
